Created a test that is functional and displays that everything works as it should
